+++ java/com/turbovnc/rfb/CSecurity.java
@@ -1,5 +1,5 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
- * Copyright (C) 2017 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2017-2018 D. R. Commander.  All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -37,6 +37,7 @@ package com.turbovnc.rfb;
 public abstract class CSecurity {
   public abstract boolean processMsg(CConnection cc);
   public abstract int getType();
+  public int getChosenType() { return getType(); }
   public abstract String getDescription();
   public abstract String getProtocol();